+++ src/Dto/BaseDto.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace ITMobile\ITMobileCommon\Dto;
+
+use Spatie\DataTransferObject\DataTransferObject;
+use Spatie\DataTransferObject\Exceptions\UnknownProperties;
+
+abstract class BaseDto extends DataTransferObject
+{
+    /**
+     * @throws UnknownProperties
+     */
+    public static function fromArray(array $data): static
+    {
+        return new static(...$data);
+    }
+}