+++ src/Dto/Company/CompanyUpdateDto.php
@@ -1,0 +1,30 @@
+<?php
+
+namespace ITMobile\ITMobileCommon\Dto\Company;
+
+use ITMobile\ITMobileCommon\Dto\BaseDto;
+use ITMobile\ITMobileCommon\Enums\Company\CompanyType;
+use Spatie\DataTransferObject\Attributes\CastWith;
+use Spatie\DataTransferObject\Casters\EnumCaster;
+
+/**
+ * @property int $id
+ * @property int|null $cityId
+ * @property string|null $name
+ * @property CompanyType|null $type
+ * @property array|null $details
+ * @property array|null $settings
+ */
+class CompanyUpdateDto extends BaseDto
+{
+    public int $id;
+
+    public ?int $cityId;
+    public ?string $name;
+
+    #[CastWith(EnumCaster::class, CompanyType::class)]
+    public ?CompanyType $type;
+
+    public ?array $details;
+    public ?array $settings;
+}